requirements.txt
pyserial==3.5
kivy==2.3.1
requests==2.32.4
